+++ docs/src/release_notes.rst
@@ -13,6 +13,7 @@ Unreleased
 - Avoid index error when handling syntax errors (:issue:`60`)
 - Construct fully-qualified names more strictly to avoid hiding other issues
   (:issue:`61`)
+- Resolve string annotations in the module scope (:issue:`62`)
 - Correctly ensure that return annotations are valid types (:issue:`63`)
 
 0.5.1 (2021-11-20)